requirements.txt
CTkMessagebox==2.0
customtkinter==5.1.2
mysql-connector-python==8.0.32
Pillow==9.5.0
pygame==2.4.0
Requests==2.31.0
winotify==1.1.0
